Move defaultConfig into constants.ts
