+++ src/constants.ts
@@ -1,6 +1,6 @@
 import { UserConfigRequired } from './types'
 
-export default {
+export const defaultConfig: UserConfigRequired = {
   createDir: true,
   dotPrefixWhitelist: {
     dirs: ['.github', '.husky', '.vscode'],
@@ -8,4 +8,4 @@ export default {
   },
   safeMode: true,
   safetyDirName: '_generated',
-} satisfies UserConfigRequired
+}
