+++ src/utils.ts
@@ -1,6 +1,6 @@
 import { existsSync, statSync } from 'fs'
 import { ParsedPath, parse } from 'path'
-import defaultConfig from './defaultConfig'
+import { defaultConfig } from './constants'
 import { Config, ConfigOrigin, UserConfig } from './types'
 
 /**
